Add github context parser to support GITHUB_ENV and GITHUB_PATH in run
