+++ nodes/gh-action@v1.go
@@ -118,15 +118,13 @@ func (n *GhActionNode) ExecuteImpl(c core.ExecutionContext) error {
 		}
 	}
 
-	// Create runner file commands
-	fileCommandUuid := uuid.New()
-	for fileCommand, envName := range contextEnvList {
-		fname := fmt.Sprintf("%s_%s", fileCommand, fileCommandUuid)
-		path := filepath.Join(sysRunnerTempDir, "_runner_file_commands", fname)
-		err := os.WriteFile(path, []byte(""), 0644)
-		if err != nil {
-			return u.Throw(err)
-		}
+	ghContextParser := GhContextParser{}
+	ctxEnvs, err := ghContextParser.Init(sysRunnerTempDir)
+	if err != nil {
+		return u.Throw(err)
+	}
+	for envName, path := range ctxEnvs {
+		// Set GITHUB_PATH, GITHUB_ENV, etc.
 		contextEnvironMap[envName] = path
 	}
 
@@ -172,47 +170,13 @@ func (n *GhActionNode) ExecuteImpl(c core.ExecutionContext) error {
 		}
 	}
 
-	githubPath := contextEnvironMap["GITHUB_PATH"]
-	// load all paths from the github path file and append them to the PATH
-	if githubPath != "" {
-		p, err := os.ReadFile(githubPath)
-		if err != nil {
-			return u.Throw(err)
-		}
-
-		newPaths := []string{}
-
-		lines := strings.Split(string(p), "\n")
-		for _, line := range lines {
-			if line == "" {
-				continue
-			}
-			newPaths = append(newPaths, line)
-		}
-
-		if len(newPaths) > 0 {
-			contextEnvironMap["PATH"] = strings.Join(newPaths, string(os.PathListSeparator)) + string(os.PathListSeparator) + contextEnvironMap["PATH"]
-		}
-
-		err = os.Remove(githubPath)
-		if err != nil {
-			return u.Throw(err)
-		}
+	// Get the context vars from GITHUB_ENV and GITHUB_PATH
+	ctxEnvs, err = ghContextParser.Parse(contextEnvironMap)
+	if err != nil {
+		return u.Throw(err)
 	}
-
-	githubEnv := contextEnvironMap["GITHUB_ENV"]
-	if githubEnv != "" {
-		b, err := os.ReadFile(githubEnv)
-		if err != nil {
-			return u.Throw(err)
-		}
-		envs, err := parseOutputFile(string(b))
-		if err != nil {
-			return u.Throw(err)
-		}
-		for envName, envValue := range envs {
-			contextEnvironMap[envName] = strings.TrimRight(envValue, "\t\n")
-		}
+	for envName, envValue := range ctxEnvs {
+		contextEnvironMap[envName] = envValue
 	}
 
 	// Transfer the output values from the github action to the node output values
@@ -306,13 +270,12 @@ func (n *GhActionNode) ExecuteDocker(c core.ExecutionContext, workingDirectory s
 		}
 	}
 
-	// Create runner file commands
+	// Update github context paths to the docker paths
 	for _, envName := range contextEnvList {
 		path := envs.ExecutionEnviron[envName]
 		if path == "" {
 			return u.Throw(fmt.Errorf("expected %s to be set in execution environment", envName))
 		}
-
 		dockerEnviron[envName] = filepath.Join(dockerGithubFileCommands, filepath.Base(path))
 	}
 
@@ -614,60 +577,6 @@ type ActionRuns struct {
 	Args  []string `json:"args"`
 }
 
-func parseOutputFile(input string) (map[string]string, error) {
-	results := make(map[string]string)
-	lines := strings.Split(input, "\n")
-
-	for i := 0; i < len(lines); i++ {
-		line := lines[i]
-		if line == "" {
-			continue
-		}
-
-		var key, value string
-		equalsIndex := strings.Index(line, "=")
-		heredocIndex := strings.Index(line, "<<")
-
-		// Normal style: NAME=VALUE
-		if equalsIndex >= 0 && (heredocIndex < 0 || equalsIndex < heredocIndex) {
-			parts := strings.SplitN(line, "=", 2)
-			if len(parts) != 2 || parts[0] == "" {
-				return nil, errors.New(fmt.Sprintf("Invalid format '%s'. Name must not be empty", line))
-			}
-			key, value = parts[0], parts[1]
-		} else if heredocIndex >= 0 && (equalsIndex < 0 || heredocIndex < equalsIndex) {
-			// Heredoc style: NAME<<EOF
-			parts := strings.SplitN(line, "<<", 2)
-			if len(parts) != 2 || parts[0] == "" || parts[1] == "" {
-				return nil, errors.New(fmt.Sprintf("Invalid format '%s'. Name and delimiter must not be empty", line))
-			}
-			key = parts[0]
-			delimiter := parts[1]
-
-			var heredocValue strings.Builder
-			for i++; i < len(lines); i++ {
-				if lines[i] == delimiter {
-					break
-				}
-				heredocValue.WriteString(lines[i])
-				if i < len(lines)-1 {
-					heredocValue.WriteString("\n")
-				}
-			}
-			if i >= len(lines) {
-				return nil, errors.New(fmt.Sprintf("Invalid value. Matching delimiter not found '%s'", delimiter))
-			}
-			value = heredocValue.String()
-		} else {
-			return nil, errors.New(fmt.Sprintf("Invalid format '%s'", line))
-		}
-
-		results[key] = value
-	}
-
-	return results, nil
-}
-
 // getRunnersDir returns the directory of the latest runner version.
 func getRunnersDir() (string, error) {
 
@@ -723,15 +632,6 @@ func sanitize(name string, allowHyphens bool) string {
 		}
 	}
 	return sb.String()
-}
-
-// https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#environment-files
-var contextEnvList = map[string]string{
-	"add_path":     "GITHUB_PATH",
-	"save_state":   "GITHUB_STATE",
-	"set_env":      "GITHUB_ENV",
-	"step_summary": "GITHUB_STEP_SUMMARY",
-	"set_output":   "GITHUB_OUTPUT",
 }
 
 // https://github.com/actions/runner/blob/f467e9e1255530d3bf2e33f580d041925ab01951/src/Runner.Worker/GitHubContext.cs#L9
