+++ nodes/run@v1.go
@@ -4,6 +4,7 @@ import (
 	"actionforge/graph-runner/core"
 	ni "actionforge/graph-runner/node_interfaces"
 	"actionforge/graph-runner/utils"
+	u "actionforge/graph-runner/utils"
 	_ "embed"
 	"errors"
 	"fmt"
@@ -59,6 +60,22 @@ func (n *RunNode) ExecuteImpl(c core.ExecutionContext) error {
 		kv := strings.SplitN(env, "=", 2)
 		if len(kv) == 2 {
 			contextEnvironMap[kv[0]] = ReplaceContextVariables(kv[1])
+		}
+	}
+
+	ghContextParser := GhContextParser{}
+	if contextEnvironMap["GITHUB_ACTIONS"] != "" {
+		sysRunnerTempDir := contextEnvironMap["RUNNER_TEMP"]
+		if sysRunnerTempDir == "" {
+			return fmt.Errorf("RUNNER_TEMP not set")
+		}
+		ctxEnvs, err := ghContextParser.Init(sysRunnerTempDir)
+		if err != nil {
+			return u.Throw(err)
+		}
+		for envName, path := range ctxEnvs {
+			// Set GITHUB_PATH, GITHUB_ENV, etc.
+			contextEnvironMap[envName] = path
 		}
 	}
 
@@ -190,6 +207,19 @@ func (n *RunNode) ExecuteImpl(c core.ExecutionContext) error {
 	}
 
 	if cmd.ProcessState.ExitCode() == 0 {
+
+		if contextEnvironMap["GITHUB_ACTIONS"] != "" {
+			// Get the context vars from GITHUB_ENV and GITHUB_PATH
+			ctxEnvs, err := ghContextParser.Parse(contextEnvironMap)
+			if err != nil {
+				return u.Throw(err)
+			}
+			for envName, envValue := range ctxEnvs {
+				contextEnvironMap[envName] = envValue
+			}
+			c.SetContextEnvironMap(contextEnvironMap)
+		}
+
 		err = n.Execute(n.Executions[ni.Run_v1_Output_exec_success], c)
 		if err != nil {
 			return err
